Update codebase for Godot 4.x compatibility
Refactored various scripts to use Godot 4.x API changes, including method and property renames (e.g., Instance -> Instantiate, GlobalTranslation -> GlobalPosition, File -> FileAccess), updated node access patterns, and replaced deprecated features. This improves compatibility and prepares the project for Godot 4.x, ensuring proper functionality and reducing runtime errors.

+++ src/health_systems/Health.cs
@@ -55,7 +55,7 @@ namespace HealthSystem
         {
             if (_regenRate > 0 && Points < PointsMax && _regenDelayTimer.IsStopped())
             {
-                IncrementPoints((float)delta * _regenRate);
+                IncrementPoints((float)(float)delta * _regenRate);
                 EmitSignal(nameof(Increased));
             }
         }
